pyproject.toml
[project]
name = "sdoptimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "itsdangerous>=2.2.0",
    "numpy>=2.0.0",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "pysd>=3.13.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
